Add Information to Models and README
Add comment to models.py: I just realized that the context links for a
given page might need an own class. This will be needed to be adressed,
when everything else works.
Add comment to README: 'project_name' needs of course to be changed when
editing files according to my notes.