Add URL, View and Template for a Single Page Content
Right now the url for the "details" view passes a "page_category_id"
which so far is not needed in the template and a page_container_id
whicht defines which object of the pag_container model is to be
desplayed.
This is not satisfying. Eventually the ID's should be replaced by
category names and page titles. Unfortunately I have no bleeding clue
how to capture these bleeding values in the urls.